+++ src/chart/options/advancedoptions.cpp
@@ -3,51 +3,6 @@
 using namespace Vizzu;
 using namespace Vizzu::Diag;
 
-OptionsSetter &ExistsHandler::addSeries(const ScaleId &scaleId,
-										 const Data::SeriesIndex &index,
-										 std::optional<size_t> pos)
-{
-	Base::addSeries(scaleId, index, pos);
-	handleExists();
-	return *this;
-}
-
-OptionsSetter &ExistsHandler::deleteSeries(const ScaleId &scaleId,
-											const Data::SeriesIndex &index)
-{
-	Base::deleteSeries(scaleId, index);
-	handleExists();
-	return *this;
-}
-
-OptionsSetter &ExistsHandler::setHorizontal(bool horizontal)
-{
-	Base::setHorizontal(horizontal);
-	handleExists();
-	return *this;
-}
-
-void ExistsHandler::handleExists()
-{
-	if (forcedExistsSeries)
-		options.getScales().visitAll(
-		[=](ScaleId id, const Scale& scale)
-		{
-			if (scale.discretesIds().empty()
-				&& scale.continousId()
-				&& scale.continousId()->getType() == Data::SeriesType::Exists)
-				Base::deleteSeries(id, Data::SeriesIndex(Data::SeriesType::Exists));
-
-			if ((Diag::isAxis(id) || id == ScaleId::size)
-				&& !scale.isEmpty()
-				&& !scale.continousId())
-			{
-				Base::addSeries(id, Data::SeriesIndex(Data::SeriesType::Exists));
-			}
-
-		});
-}
-
 OptionsSetter &AdvancedOptions::deleteSeries(const ScaleId &scaleId, const Data::SeriesIndex &index)
 {
 	Base::deleteSeries(scaleId, index);