+++ src/data/datacube/aggregator.h
@@ -13,7 +13,7 @@ namespace Data
 class Aggregator
 {
 public:
-	enum Type : uint32_t { Exists, Sum, Count, Distinct, Min, Max, Mean };
+	enum Type : uint32_t { Sum, Count, Distinct, Min, Max, Mean };
 
 	explicit Aggregator(Type type);
 	Aggregator &add(double, int);