+++ src/data/datacube/seriestype.cpp
@@ -4,7 +4,6 @@ using namespace Vizzu::Data;
 
 const SeriesType SeriesType::Index		= SeriesType( false, CT::Discrete, 1u, "index" );
 const SeriesType SeriesType::Discrete	= SeriesType( true,  CT::Discrete, 2u, "discrete" );
-const SeriesType SeriesType::Exists		= SeriesType( false, CT::Continous, AT::Exists, "exists" );
 const SeriesType SeriesType::Sum		= SeriesType( true,  CT::Continous, AT::Sum, "sum");
 const SeriesType SeriesType::Count		= SeriesType( false, CT::Continous, AT::Count, "count" );
 const SeriesType SeriesType::Min		= SeriesType( true,  CT::Continous, AT::Min, "min" );
@@ -16,7 +15,6 @@ const std::vector<SeriesType> SeriesType::constTypes
 {
 	SeriesType::Index,
 	SeriesType::Discrete,
-	SeriesType::Exists,
 	SeriesType::Sum,
 	SeriesType::Count,
 	SeriesType::Min,